refactor: chat featureにrepositoryレイヤーを追加
- chat-usage-repository.tsを新規作成し、chat_usage_eventsテーブルへのクエリを集約
- cost-tracker.tsからSupabase直接呼び出しを削除し、repository関数を利用
